+++ components/frontend/src/AppUI.test.jsx
@@ -7,7 +7,14 @@ import { dataModel, report } from "./__fixtures__/fixtures"
 import * as fetchServerApi from "./api/fetch_server_api"
 import { AppUI } from "./AppUI"
 import { mockGetAnimations } from "./dashboard/MockAnimations"
-import { asyncClickText, clickText, expectNoAccessibilityViolations, expectSearch, expectText } from "./testUtils"
+import {
+    asyncClickText,
+    clickText,
+    expectNoAccessibilityViolations,
+    expectNoText,
+    expectSearch,
+    expectText,
+} from "./testUtils"
 import { theme } from "./theme"
 
 beforeEach(async () => {
@@ -20,7 +27,7 @@ beforeEach(async () => {
 
 afterEach(() => vi.restoreAllMocks())
 
-async function renderAppUI(reports) {
+async function renderAppUI({ reports = [], reportDate = null, reportUuid = "" } = {}) {
     let result
     await act(async () => {
         result = render(
@@ -29,8 +36,8 @@ async function renderAppUI(reports) {
                     dataModel={dataModel}
                     handleDateChange={vi.fn}
                     lastUpdate={new Date()}
-                    reportDate={reports ? null : undefined}
-                    reportUuid={reports ? "report_uuid" : ""}
+                    reportDate={reportDate}
+                    reportUuid={reportUuid}
                     reports={reports ?? []}
                     reportsOverview={{}}
                     user="xxx"
@@ -42,13 +49,13 @@ async function renderAppUI(reports) {
 }
 
 it("shows an error message when there are no reports", async () => {
-    const { container } = await renderAppUI()
+    const { container } = await renderAppUI({ reportDate: new Date() })
     expectText(/Sorry, no reports/)
     await expectNoAccessibilityViolations(container)
 })
 
 it("handles sorting", async () => {
-    await renderAppUI([report])
+    await renderAppUI({ reports: [report], reportUuid: "report_uuid" })
     clickText("Comment", 0)
     expectSearch("?sort_column_report_uuid=comment")
     clickText("Status", 0)
@@ -66,8 +73,73 @@ it("handles sorting", async () => {
 
 it("resets all settings", async () => {
     history.push("?date_interval=2")
-    const { container } = await renderAppUI()
-    await expectNoAccessibilityViolations(container)
+    await act(async () => {
+        await renderAppUI()
+    })
     clickText("Reset settings")
     expectSearch("")
+})
+
+it("shows all tags of all reports in the settings menu at the reports overview", async () => {
+    await renderAppUI({
+        reports: [
+            {
+                report_uuid: "report_uuid1",
+                subjects: { subject_uuid1: { metrics: { metric_uuid1: { scale: "count", tags: ["tag1"] } } } },
+                title: "Report 1",
+            },
+            {
+                report_uuid: "report_uuid2",
+                subjects: { subject_uuid2: { metrics: { metric_uuid2: { scale: "count", tags: ["tag2"] } } } },
+                title: "Report 2",
+            },
+        ],
+    })
+    expectText("tag1")
+    expectText("tag2")
+})
+
+it("shows only tags of the current report in the settings menu", async () => {
+    await renderAppUI({
+        reports: [
+            {
+                report_uuid: "report_uuid1",
+                subjects: {
+                    subject_uuid1: {
+                        metrics: {
+                            metric_uuid1: {
+                                direction: "<",
+                                name: "Metric 1",
+                                scale: "count",
+                                tags: ["tag1"],
+                                unit: "unit",
+                            },
+                        },
+                    },
+                },
+                title: "Report 1",
+            },
+            {
+                report_uuid: "report_uuid2",
+                subjects: {
+                    subject_uuid2: {
+                        metrics: {
+                            metric_uuid2: {
+                                direction: "<",
+                                name: "Metric 2",
+                                scale: "count",
+                                tags: ["tag2"],
+                                unit: "unit",
+                            },
+                        },
+                    },
+                },
+                title: "Report 2",
+            },
+        ],
+        reportUuid: "report_uuid1",
+    })
+    clickText("Settings")
+    expectText("tag1", 3)
+    expectNoText("tag2")
 })