+++ docs/src/changelog.md
@@ -18,6 +18,7 @@ If your currently installed *Quality-time* version is not the penultimate versio
 
 - Fix broken links in the documentation. Fixes [#12033](https://github.com/ICTU/quality-time/issues/12033).
 - Disable issue tracker fields until the user picks an issue tracker type. Fixes [#12039](https://github.com/ICTU/quality-time/issues/12039).
+- The "Visible tags" menu in the settings panel of a report would show tags not used in the current report. Fixes [#12061](https://github.com/ICTU/quality-time/issues/12061).
 
 ### Changed
 
@@ -25,8 +26,8 @@ If your currently installed *Quality-time* version is not the penultimate versio
 
 ### Added
 
-- Metrics can have an optional secondary name. Closes [#12023](https://github.com/ICTU/quality-time/issues/12023).
 - When measuring time remaining with Jira as source, add a parameter to select the event until which to measure the time remaining so it is clear what time is being measured. Only value supported at the moment is the end of the active sprint. Closes [#11893](https://github.com/ICTU/quality-time/issues/11893).
+- Metrics can have an optional secondary name. Closes [#12023](https://github.com/ICTU/quality-time/issues/12023).
 
 ## v5.43.0 - 2025-09-26